Read modbus command

+++ main.c
@@ -86,7 +86,7 @@ void mgmt_coil_register(uint8_t *modbus_msg) {
 
     if ((((modbus_msg[4] << 8) | modbus_msg[5]) >= ADDRESS_COIL_REGISTER_0) && (((modbus_msg[4] << 8) | modbus_msg[5]) <= ADDRESS_COIL_REGISTER_3)) {
 
-        switch (((modbus_msg[2] << 8) | modbus_msg[3])) {
+        switch ((uint16_t)((modbus_msg[2] << 8) | modbus_msg[3])) {
             case COIL_ON:
                 set_gpio_high_level(&PORTC, (modbus_msg[4] << 8) | modbus_msg[5]);
             break;
